+++ chevah/utils/tests/normal/system/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2012 Adi Roiban.
+# See LICENSE for details.
+"""
+System test running under non-privileged account.
+"""