+++ release-notes.rst
@@ -1,6 +1,11 @@
 Release notes for chevah.utils
 ==============================
 
+0.20.2 - 30/07/2013
+-------------------
+
+* Merge latest fixed from 0.19.4.
+
 
 0.20.1 - 21/07/2013
 -------------------
@@ -13,6 +18,19 @@ Release notes for chevah.utils
 
 * Don't set invalid values for configuration options and raise an error when
   trying to set one.
+
+
+0.19.4 - 22/07/2013
+-------------------
+
+* Removed redundant code from the message when generating SSH key.
+
+
+0.19.3 - 19/07/2013
+-------------------
+
+* Added a more user friendly message when generating SSH key and no
+  comment is specified.
 
 
 0.19.2 - 18/07/2013